ported update_slice from swift binding
